+++ cocos/physics/CCPhysicsContact.h
@@ -25,6 +25,7 @@
 #ifndef __CCPHYSICS_CONTACT_H__
 #define __CCPHYSICS_CONTACT_H__
 
+#include "ccConfig.h"
 #ifdef CC_USE_PHYSICS
 
 #include "CCObject.h"
@@ -44,7 +45,7 @@ typedef Point Vect;
 
 typedef struct PhysicsContactData
 {
-    static const long POINT_MAX = 4;
+    static const int POINT_MAX = 4;
     Point points[POINT_MAX];
     int   count;
     Point normal;
