+++ cocos/physics/CCPhysicsShape.h
@@ -25,6 +25,7 @@
 #ifndef __CCPHYSICS_SHAPE_H__
 #define __CCPHYSICS_SHAPE_H__
 
+#include "ccConfig.h"
 #ifdef CC_USE_PHYSICS
 
 #include "CCObject.h"
@@ -206,7 +207,7 @@ public:
     
     Point getPoint(int i) const;
     void getPoints(Point* outPoints) const;
-    long getPointsCount() const;
+    int getPointsCount() const;
     virtual Point getCenter() override;
 protected:
     bool init(const Point* points, int count, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, const Point& offset = Point::ZERO);
@@ -249,7 +250,7 @@ public:
     static PhysicsShapeEdgeBox* create(const Size& size, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 0, const Point& offset = Point::ZERO);
     virtual Point getOffset() override { return _offset; }
     void getPoints(const Point* outPoints) const;
-    long getPointsCount() const;
+    int getPointsCount() const;
     
 protected:
     bool init(const Size& size, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 1, const Point& offset = Point::ZERO);
@@ -271,7 +272,7 @@ public:
     static PhysicsShapeEdgePolygon* create(const Point* points, int count, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 1);
     virtual Point getCenter() override;
     void getPoints(Point* outPoints) const;
-    long getPointsCount() const;
+    int getPointsCount() const;
     
 protected:
     bool init(const Point* points, int count, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 1);
@@ -293,7 +294,7 @@ public:
     static PhysicsShapeEdgeChain* create(const Point* points, int count, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 1);
     virtual Point getCenter() override;
     void getPoints(Point* outPoints) const;
-    long getPointsCount() const;
+    int getPointsCount() const;
     
 protected:
     bool init(const Point* points, int count, const PhysicsMaterial& material = PHYSICSSHAPE_MATERIAL_DEFAULT, float border = 1);
